+++ lib/apis/sign_api/models/session_models.dart
@@ -25,6 +25,11 @@ class SessionProposalCompleter {
     required this.completer,
     this.sessionProperties,
   });
+
+  @override
+  String toString() {
+    return 'SessionProposalCompleter(id: $id, selfPublicKey: $selfPublicKey, pairingTopic: $pairingTopic, requiredNamespaces: $requiredNamespaces, optionalNamespaces: $optionalNamespaces, sessionProperties: $sessionProperties, completer: $completer)';
+  }
 }
 
 @JsonSerializable()
@@ -43,6 +48,11 @@ class Namespace {
       _$NamespaceFromJson(json);
 
   Map<String, dynamic> toJson() => _$NamespaceToJson(this);
+
+  @override
+  String toString() {
+    return 'Namespace(accounts: $accounts, methods: $methods, events: $events)';
+  }
 
   @override
   bool operator ==(Object other) {
@@ -68,6 +78,7 @@ class Namespace {
 @JsonSerializable(includeIfNull: false)
 class SessionData {
   final String topic;
+  final String pairingTopic;
   final Relay relay;
   int expiry;
   bool acknowledged;
@@ -81,6 +92,7 @@ class SessionData {
 
   SessionData({
     required this.topic,
+    required this.pairingTopic,
     required this.relay,
     required this.expiry,
     required this.acknowledged,
@@ -97,6 +109,11 @@ class SessionData {
       _$SessionDataFromJson(json);
 
   Map<String, dynamic> toJson() => _$SessionDataToJson(this);
+
+  @override
+  String toString() {
+    return 'SessionData(topic: $topic, pairingTopic: $pairingTopic, relay: $relay, expiry: $expiry, acknowledged: $acknowledged, controller: $controller, namespaces: $namespaces, requiredNamespaces: $requiredNamespaces, optionalNamespaces: $optionalNamespaces, sessionProperties: $sessionProperties, self: $self, peer: $peer)';
+  }
 }
 
 @JsonSerializable()
@@ -120,4 +137,9 @@ class SessionRequest {
       _$SessionRequestFromJson(json);
 
   Map<String, dynamic> toJson() => _$SessionRequestToJson(this);
+
+  @override
+  String toString() {
+    return 'SessionRequest(id: $id, topic: $topic, method: $method, chainId: $chainId, params: $params)';
+  }
 }